Move wait start to power board class
